# xiaoyang-web-http-pb-152026/docs/apis.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "xiaoyang api docs",
    "version": "2.0"
  },
...
  "host": "localhost:8080",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/auth/login": {
      "post": {
        "summary": "登录",
        "description": "登录",
        "operationId": "UserService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LoginRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "summary": "登出",
        "description": "登出",
        "operationId": "UserService_Logout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1LogoutResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LogoutRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/auth/register": {
      "post": {
        "summary": "注册",
        "description": "注册",
        "operationId": "UserService_Register",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1RegisterResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RegisterRequest"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
...
    "/v1/user/password/{qqNumber}": {
      "put": {
        "summary": "修改密码",
        "description": "修改密码",
        "operationId": "UserService_ChangePassword",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ChangePasswordResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "qqNumber",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserServiceChangePasswordBody"
            }
          }
        ],
        "tags": [
          "UserService"
        ]
      }
    },
    "/v1/user/{qqNumber}": {
      "get": {
        "summary": "获取用户信息",
        "description": "获取用户信息",
        "operationId": "UserService_GetUserInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetUserInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "qqNumber",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "UserService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "summary": "更新用户信息",
        "description": "更新用户信息",
        "operationId": "UserService_UpdateUserInfo",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateUserInfoResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "qqNumber",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserServiceUpdateUserInfoBody"
            }
          }
        ],
        "tags": [
          "UserService"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
...
    "UserServiceChangePasswordBody": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "oldPassword": {
          "type": "string"
        },
        "newPassword": {
          "type": "string"
        }
      },
      "title": "修改密码请求"
    },
    "UserServiceUpdateUserInfoBody": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "nickname": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      },
      "title": "更新用户信息请求"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1ChangePasswordResponse": {
      "type": "object",
      "title": "修改密码响应"
    },
...
    "v1GetUserInfoResponse": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "qqNumber": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      },
      "title": "获取用户信息响应"
    },
...
    "v1LoginRequest": {
      "type": "object",
      "properties": {
        "qqNumber": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "title": "登录请求"
    },
    "v1LoginResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "qqNumber": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      },
      "title": "登录响应"
    },
    "v1LogoutRequest": {
      "type": "object",
      "properties": {
        "qqNumber": {
          "type": "string"
        }
      },
      "title": "登出请求"
    },
    "v1LogoutResponse": {
      "type": "object",
      "title": "登出响应"
    },
    "v1RegisterRequest": {
      "type": "object",
      "properties": {
        "qqNumber": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        }
      },
      "title": "注册请求"
    },
    "v1RegisterResponse": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "userId": {
          "type": "integer",
          "format": "int64"
        },
        "qqNumber": {
          "type": "string"
        },
        "nickname": {
          "type": "string"
        },
        "avatar": {
          "type": "string"
        }
      },
      "title": "注册响应"
    },
...
    "v1UpdateUserInfoResponse": {
      "type": "object",
      "title": "更新用户信息响应"
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "description": "Input a \"Bearer your-jwt-token\" to Value",
      "name": "Authorization",
      "in": "header"
    },
    "LeagueAuth": {
      "type": "apiKey",
      "description": "Input league key to Value Input league key to Value",
      "name": "LeagueKey",
      "in": "header"
    }
  }
}